+++ src/pygubudesigner/preferences.py
@@ -1,6 +1,7 @@
 import configparser
 import logging
 import os
+import json
 import tkinter as tk
 from pathlib import Path
 from tkinter import filedialog, messagebox
@@ -8,7 +9,7 @@ from tkinter import filedialog, messagebox
 import pygubu
 from appdirs import AppDirs
 
-from pygubudesigner.util import get_ttk_style
+from pygubudesigner.services.theming import get_ttk_style
 
 from .i18n import translator as _
 
@@ -54,6 +55,9 @@ options = {
         "default": "no",
     },
     "v_style_definition_file": {
+        "default": "",
+    },
+    "maindock_layout": {
         "default": "",
     },
 }
@@ -143,6 +147,10 @@ def get_window_size():
     return get_option("geometry")
 
 
+def get_preview_indicator_color():
+    return "red"
+
+
 def save_from_dict(new_values: dict):
     # General
     for key, value in new_values.items():
@@ -150,6 +158,21 @@ def save_from_dict(new_values: dict):
     # Custom Widgets
     config.remove_section(SEC_CUSTOM_WIDGETS)
     save_configfile()
+
+
+def save_maindock_layout(layout: dict):
+    value = json.dumps(layout)
+    set_option("maindock_layout", value, True)
+
+
+def get_maindock_layout():
+    layout = {}
+    try:
+        jvalue = get_option("maindock_layout")
+        layout = json.loads(jvalue)
+    except json.JSONDecodeError:
+        pass
+    return layout
 
 
 # Get user configuration