+++ src/pygubudesigner/services/fieldvalidator.py
@@ -2,11 +2,12 @@ import keyword
 import pathlib
 
 from pygubu.forms.validation.base import Constraint, ConstraintValidator
+from pygubu.i18n import noop_translator as _
 
 
 class IsIdentifier(Constraint):
     code = "not_identifier_error"
-    message = "This value should be a valid python identifier."
+    message = _("This value should be a valid python identifier.")
 
     def __init__(self, *args, message=None, **kw):
         super().__init__(*args, **kw)
@@ -29,9 +30,34 @@ class IsIdentifierValidator(ConstraintValidator):
         )
 
 
+class Choice(Constraint):
+    code = "invalid_choice"
+    message = _("Select a valid choice.")
+
+    def __init__(self, *args, choices=None, message=None, **kw):
+        super().__init__(*args, **kw)
+        if message is not None:
+            self.message = message
+        self.valid_choices = choices if choices is not None else []
+
+    def validated_by(self):
+        return ChoiceValidator
+
+
+class ChoiceValidator(ConstraintValidator):
+    def validate(self, value, constraint):
+        if value not in constraint.valid_choices:
+            self.context.add_violation(
+                message=constraint.message,
+                constraint=constraint,
+                code=constraint.code,
+                params=None,
+            )
+
+
 class PathExist(Constraint):
     code = "path_not_exits_error"
-    message = "The path must exist."
+    message = _("The path must exist.")
 
     def __init__(self, *args, message=None, **kw):
         super().__init__(*args, **kw)